pyproject.toml
[project]
name = "designer"
version = "0.1.0"
description = "AI-powered visual content generator for Instagram carousels and ads"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.89.0",
    "tavily-python>=0.7.23",
    "pytrends>=4.9.2",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "moviepy>=2.2.1",
    "google-auth-oauthlib>=1.3.1",
    "google-generativeai>=0.8.6",
    "google-genai>=1.70.0",
    "streamlit>=1.56.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["designer"]
